DBZ-7024 Move reuseable testing functions for DebeziumEngine into common util class
